+++ package.json
@@ -8,7 +8,7 @@
     "url": "ssh://git@github.com:storatech/stora-utils.git",
     "directory": "@storatech/stora-utils"
   },
-  "version": "1.1.6",
+  "version": "1.1.7",
   "description": "stora common utils",
   "scripts": {
     "build": "rm -rf ./dist && tsc -p . && cp package.json ./dist",
